# src/main/java/com/flash_loan/bank/account_service/infrastructure/adapters/in/web/dto/request/AccountBalanceOperationRequestDto.java
package com.flash_loan.bank.account_service.infrastructure.adapters.in.web.dto.request;

import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import lombok.Data;

import java.math.BigDecimal;

@Data
public class AccountBalanceOperationRequestDto {

    @NotNull
    @DecimalMin(value = "0.01", inclusive = true)
    private BigDecimal amount;

    @NotBlank
    private String transactionId;
}